# Projeto1/swagger.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Projeto1",
        "description": "Projeto1 Back_End",
        "contact":{
            "name": "TPSI-DWB"
        },
        "servers":[
            "http://localhost:8000"
        ]
    },
    "paths":{
        "definitions":{
            "cars": {
                "type": "object",
                "properties" : {
                    "id": {
                        "primaryKey": true,
                        "type": "integer",
                        "autoIncrement": true,
                        "allowNull": false
                    },
                    "seller_id": {
                        "type": "integer",
                        "allowNull": false
                    },
                    "brandmodel": {
                        "type":"string",
                        "allowNull": false
                    }, 
                    "description": {
                        "type":"string",
                        "allowNull": false
                    },
                    "price": {
                        "type":"decimal",
                        "allowNull": false
                    },
                    "url": {
                        "type":"string",
                        "allowNull": false
                    },
                    "views": {
                        "type":"integer",
                        "format":"int64",
                        "allowNull": false
                    },
                    "images": {
                        "type":"json",
                        "allowNull": false
                    },
                    "comments": {
                        "type":"json",
                        "allowNull": false
                    },
                    "tags": {
                        "type":"json",
                        "allowNull": false
                    }

                }
            }
        },
        
        "/":{
            "get":{
                "tags":[
                    "/"
                ],
                "summary":"Mostrar todos os carros",
                "description":"Listar carros existentes na base de dados",
                "operationId":"mostracarros",
                "produces":[
                    "application/xml",
                    "application/json"
                ],
                "parameters":[
                    {
                        "in":"body",
                        "name":"body",
                        "description":"Mostrar carros",
                        "required":true,
                        "schema":{
                            "$ref":"#/definitions/cars"
                        }
                    }
                ],
                "responses":{
                    "default":{
                        "description":"Operação bem sucedida"
                    }
                }
            }
        },

        "/novo":{
            "post":{
                "tags":[
                    "novo"
                ],
                "summary":"Adiciona carro",
                "description":"Adiciona um Carro na base de dados",
                "operationId":"AddCar",
                "produces":[
                    "application/xml",
                    "application/json"
                ],
                "parameters":[
                    {
                        "name":"body",
                        "in":"body",
                        "description":"Person id",
                        "required":true,
                        "type":"string"
                    }
                ],
                "responses":{
                    "200":{
                        "description":"Operação bem sucedida",
                        "schema":{
                            "$ref":"#/definitions/Person"
                        }
                    }
                }
            }
        },

        "/seller":{
            "get":{
                "tags":[
                    "seller"
                ],
                "summary":"Mostrar carros pelo seller_id",
                "description":"Mostrar carros pertencentes a um seller",
                "operationId":"mostracarrosseller",
                "produces":[
                    "application/xml",
                    "application/json"
                ],
                "responses":{
                    "200":{
                        "description":"Operação bem sucedida",
                        "schema":{
                            "$ref":"#/definitions/Person"
                        }
                    },
                    "400":{
                        "description":"Invalid id supplied"
                    },
                    "404":{
                        "description":"Id not found"
                    }
                }
            }
        },

        "/update/:id/": {
            "post": {
                "tags": [
                    "update/:id"
                ],
                "summary": "Adicionar views",
                "description": "Adiciona mais uma views e atualiza a tabela",
                "operationId": "addviewsid",
                "produces": [
                    "application/xml",
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Operação bem sucedida",
                        "schema": {
                            "$ref": "#/definitions/cars"
                        }
                    },
                    "400": {
                        "description": "Invalid id supplied"
                    },
                    "404": {
                        "description": "Id not found"
                    }
                }
            }
        },

        "/tag": {
            "get": {
                "tags": [
                    "tag"
                ],
                "summary": "Lista todos os carros que tenham determinados tags",
                "description": "Lista todos os carros que tenham determinados tags e devolve a lista na resposta",
                "operationId": "postag",
                "produces": [
                    "application/xml",
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Operação bem sucedida",
                        "schema": {
                            "$ref": "#/definitions/cars"
                        }
                    },
                    "400": {
                        "description": "Invalid id supplied"
                    },
                    "404": {
                        "description": "Id not found"
                    }
                }
            }
        },

        "/carroid": {
            "get": {
                "tags": [
                    "carroid"
                ],
                "summary": "Seleciona apenas um carro pelo ID",
                "description": "Seleciona apenas um carro pelo ID e devolve-o na resposta. ",
                "operationId": "getcarsid",
                "produces": [
                    "application/xml",
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Operação bem sucedida",
                        "schema": {
                            "$ref": "#/definitions/cars"
                        }
                    },
                    "400": {
                        "description": "Invalid id supplied"
                    },
                    "404": {
                        "description": "Id not found"
                    }
                }
            }
        },

        "/cars/:id": {
            "delete": {
                "tags": [
                    "cars/:id"
                ],
                "summary": "Apaga um carro existente",
                "description": "Apaga um carro existente e devolve o número de linhas atualizadas",
                "operationId": "deletecarsapagarid",
                "produces": [
                    "application/xml",
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Operação bem sucedida",
                        "schema": {
                            "$ref": "#/definitions/cars"
                        }
                    },
                    "400": {
                        "description": "Invalid id supplied"
                    },
                    "404": {
                        "description": "Id not found"
                    }
                }
            }
        },

        "/adicionaimg/:id": {
            "put": {
                "tags": [
                    "/adicionaimg/:id"
                ],
                "summary": "Adiciona o caminho de imagens para um determinado carro pelo seu ID",
                "description": "Adiciona o caminho de imagens para um determinado carro pelo seu ID e devolve a entrada atualizada na resposta. ",
                "operationId": "putcarrosimagemid",
                "produces": [
                    "application/xml",
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Operação bem sucedida",
                        "schema": {
                            "$ref": "#/definitions/cars"
                        }
                    },
                    "400": {
                        "description": "Invalid id supplied"
                    },
                    "404": {
                        "description": "Id not found"
                    }
                }
            }
        },

        "/adicionacmt": {
            "post": {
                "tags": [
                    "adicionacmt"
                ],
                "summary": "Adiciona comentario para um carro pelo seu ID",
                "description": "Adiciona comentario para um carro pelo seu ID e manda comentario via body. ",
                "operationId": "putcarrosimagemid",
                "produces": [
                    "application/xml",
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Operação bem sucedida",
                        "schema": {
                            "$ref": "#/definitions/cars"
                        }
                    },
                    "400": {
                        "description": "Invalid id supplied"
                    },
                    "404": {
                        "description": "Id not found"
                    }
                }
            }
        },

        "/listaview": {
            "get": {
                "tags": [
                    "listaview"
                ],
                "summary": "Lista todos os carros ordenados por ordem de views",
                "description": "Lista todos os carros ordenados por ordem de views e devolve a lista ordenada na resposta",
                "operationId": "getviewsorder",
                "produces": [
                    "application/xml",
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Operação bem sucedida",
                        "schema": {
                            "$ref": "#/definitions/cars"
                        }
                    },
                    "400": {
                        "description": "Invalid id supplied"
                    },
                    "404": {
                        "description": "Id not found"
                    }
                }
            }
        }
    }
}